chore(ingestion): extract shared getRetryAttempts test helper
The helper for reading the retry-attempts histogram was duplicated verbatim
in both retry test files. Move it to a sibling metrics.test-utils.ts so a
change to the metric name or how prom-client exposes _count/_sum only needs
updating in one place.
